moved linalg.py
