+++ dictos/calculus/finite_difference.py
@@ -12,7 +12,7 @@ from dictos.utils import (
     extract_coefficients_as_numer_denom,
     sort_by_subscript,
 )
-from dictos.linalg import dot_product, div
+from dictos.linalg.linalg import dot_product, div
 from dictos.poly.lagrangian_polynomial import lagrangian_poly, derivative
 from dictos.taylor_expansion import taylor_series, derivative_symbol
 from dictos.error.finite_difference import UnsupportedOrderOfDerivativeError
