+++ dictos/linalg/linalg.py
